添加FocusService.cs
