+++ Kino.Toolkit.Wpf/Kino.Toolkit.Wpf.Samples/Samples/Form/FormSample.xaml
@@ -27,14 +27,14 @@
         <Style TargetType="kino:KinoForm">
             <Setter Property="ScrollViewer.VerticalScrollBarVisibility"
                     Value="Disabled" />
-            <Setter Property="kino:FrameworkElementService.IsAutoFocus"
+            <Setter Property="kino:FocusService.IsAutoFocus"
                     Value="False" />
         </Style>
     </local:SamplePage.Resources>
     <StackPanel Background="White">
         <local:SampleSection Header="使用Form">
             <kino:KinoForm Width="500"
-                           kino:FrameworkElementService.IsAutoFocus="True"
+                           kino:FocusService.IsAutoFocus="True"
                            HorizontalAlignment="Left"
                            Header="NormalForm">
                 <kino:KinoFormTitle Content="用户信息" />
